feat: useUserName 추가 및 적용 #154
